// src/types/weather/airPollutionData.ts
export interface AirPollutionData {
  coord: {
    lat: number;
    lon: number;
  };
  list: {
    dt: number;
    main: {
      aqi: number;
    };
    components: {
      co: number;
      no: number;
      no2: number;
      o3: number;
      so2: number;
      pm2_5: number;
      pm10: number;
      nh3: number;
    };
  }[];
}